+++ lib/oli_web/controllers/api/attempt_controller.ex
@@ -408,7 +408,7 @@ defmodule OliWeb.Api.AttemptController do
         "section_slug" => section_slug,
         "activity_attempt_guid" => activity_attempt_guid,
         "part_attempt_guid" => attempt_guid,
-        "input" => input
+        "response" => input
       }) do
     datashop_session_id = Plug.Conn.get_session(conn, :datashop_session_id)
 
@@ -416,7 +416,13 @@ defmodule OliWeb.Api.AttemptController do
            section_slug,
            activity_attempt_guid,
            [
-             %{attempt_guid: attempt_guid, input: input}
+             %{
+               attempt_guid: attempt_guid,
+               input: %StudentInput{
+                 files: Map.get(input, "files", []),
+                 input: Map.get(input, "input")
+               }
+             }
            ],
            datashop_session_id
          ) do
@@ -437,8 +443,20 @@ defmodule OliWeb.Api.AttemptController do
        responses: %{
          200 => {"Evaluation response", "application/json", UserStateUpdateResponse}
        }
-  def new_part(conn, %{"activity_attempt_guid" => _, "part_attempt_guid" => _attempt_guid}) do
-    json(conn, %{"type" => "success"})
+  def new_part(conn, %{
+        "activity_attempt_guid" => activity_attempt_guid,
+        "part_attempt_guid" => part_attempt_guid
+      }) do
+    datashop_session_id = Plug.Conn.get_session(conn, :datashop_session_id)
+
+    case Activity.reset_part(activity_attempt_guid, part_attempt_guid, datashop_session_id) do
+      {:ok, attempt_state} ->
+        json(conn, %{"type" => "success", "attemptState" => attempt_state})
+
+      {:error, e} ->
+        {_, msg} = Oli.Utils.log_error("Could not reset part", e)
+        error(conn, 500, msg)
+    end
   end
 
   @doc """